Extract implementation to the source code
